add possibility to edit array of struct
it's now possible to change the name, type and size_unit and nb_unit
of a array of struct. The size_unit and nb_unit is check to avoid
negative or null size and overlapping member.
Change in size_unit result in a change in the size of the inner struct,
the inner struct have padding add is the necessary and member remove
in case of a reduction of size