Add support for multi-process program
Change the creation of the name of output json file
to be able to handle forking programe.
Now it's possible to chose a name only for the first file created
(to simplify usage on non forking program).
If a file already with this name (or it's a fork) a default name
will be created for the output file.
Default name are <prog_name>.<pid>