divide per 2 nb alloc for block/access/orig
not real perf improvment noticed but this is a necessary step to be
able to use more 2go of ram in dynStruct.
Next step will be to change global_alloc by costum alloc to alloc
memory per page for access and orig.

Also remove a memory leak on the creation of new orig.