Change alloc method for block/access/orig and display free_by_realloc
block are now alloc with dr_custom_malloc without any flag (alloc on
the heap but non code cache reachable). This is suppose to remove the
limit of 2Go of ram.
access and orig are alloc via dr_custom_malloc with non_heap flag,
this allow directly page of memory and the pages are used as array
of struct. This is also suppose to remove the memory limitation,
reduce the overhead of the managed heap and really free pages after
bock flushing.

In the interface if a block was free by realloc it was display as
never free, now it's fixed.

The memory overhead is a lot reduce with that, perf are a little bit
better.

But the limit of 2Go of ram is still present when trying to run malloc