Change allocation for orig/access to non_dr
Previous changes on allocation system was always used the heap managed
by DynamoRIO and always need to be cache-recheable because of
unimplemented feature. So the limitation du to the cache-recheable
memory was still present.
NON_DR memory is not hidden for the program, so the actual solution
is not transparent, but work better it remove the limitation of the
previous one.

When the possibility to allocate hidden memory pages will be add to
DR, dynStruct will use it.