Add recording af context instruction
The instruction which does the memory access if usually a MOV.
This doesn't provide a lot of additionnal information.
So to have more information for the structure recovery a context
instruction is now recorded. The context instruction is the next
instruction in the case of a read access and the previous one in
the case of a write access.
Sometime no context instruction will be available (because the
access is at the start or the end of a basic block).

The context instruction is not yet display in the web_ui, that will
be in the next commit.

Also add a new idea in todo/future_work which is replace the context
instruction by a little bit of static analysis around the instructio
on the binary itself. But this will not be done soon.