Display context isntruction in web_ui
The display insturciton which will be used to recover type
is now display in the web_ui. They are prefix by prev or next
to indicate theire positions relative to the accessing instructions.
If no context instruction are available 'No context' is displayed.

Fix json output for context instruction opcode.
The context instruction for read access was the preivous one, it's
now change to the next one.