Opti of the structure recovering
Now structure recovering is more quickly, because the recovering
of type and size si done only time even if it's a new struct.
Also type are now handle correctly during the comparaison between
block en existing struct.
A block is consider as struct if the only diff are some member
replace by padding, in that case the struct are merge after.
Array detection handle type correctly as well