Fix access search by offset in member detailed view and better recovery
The access search by offset was handling the update of the offset
relative to the start of the actual member.
Also in the analysis the type double is now possible only for a
size of 8 byte.
In the structure recovery, a block with 1 or multiple arrays or
simple member all of the same type and (or not) padding is not
consider as a structure