Fix print ptr make header compilable + merge struct if pointer differ
The print for ptr with comment and function ptr was not compilable.
When the only diff between 2 struct is a standar ptr vs a beter ptr
(func_ptr or ptr with comment) the struct are merge keeping the
most complete type.
Also when the only diff is a ptr of any kind vs basic type for ptr
size, the struct are merge as well keeping the ptr as type for the
member