add possibility to run recovery for 1 specific block via the web_ui
WHen the option -k is not use, all struct looking like an array are
removed, also a structure can be erase by accident by the user.
These are the reason fo the add of the "analyze this block" addition,
this link will perform the same analyze than at the start of
dynStruct.py and will not remove the structure even if it look like
an array.
This can also be useful with -n option to analyze only the interesting
block

Also update future_work.md with the last idea/features add to
dynStruct