Fix -a ld -w ld
Now we can monitor ld using option '-a ld -w ld' without segf
it of course also possible to monitor it via '-m ld'.
The fix remove the possibility to have to time the same block
in the tree (even if malloc is called multiple time for this block)
This also allow to remove the ugly trix for the init of
malloc/realloc in the libc, making it more portable for alternate
libc.