Add buffering for json output.
The buffering is done via a macro to not have to use
vprintf family function to not expose va_list.
The buffer is of 4 memory pages and the maximum write in 1 time
is 1 memory pages. If a write is longer than a memory page dynStruct
will not segfault but the a part of the output will be lost.
But this is unlikely to happen.